feat: skeleton for pool tables
