+++ apps/hubble-stats/data/index.ts
@@ -1,16 +1,8 @@
 export { default as getPoolMetadataData } from './getPoolMetadataData';
 export { default as getPoolOverviewCardData } from './getPoolOverviewCardData';
 export {
-  default as getPoolOverviewTableData,
-  type PoolOverviewTableDataType,
-} from './getPoolOverviewTableData';
-export {
   default as getPoolTransactionsData,
   type PoolTransactionDataType,
 } from './getPoolTransactionsData';
-export {
-  default as getPoolWrappingTableData,
-  type PoolWrappingTableDataType,
-} from './getPoolWrappingTableData';
 export { default as getTvlChartData } from './charts/getOverviewTvlChartData';
 export { default as getVolumeChartData } from './charts/getOverviewVolumeChartData';
